Fix all 116 implicit double-to-float conversion warnings
Fixed implicit conversion warnings by adding 'f' suffix to float literals
and explicit casts where necessary.

Changes:
- src/viola/glib_x.c: Added 'f' suffix to GLShadeColor() parameters
- src/viola/gram.y: Changed -1.0 to -1.0f in float operations
- src/viola/violalex.c: Added explicit cast for atof() result
- src/viola/htmath.c: Added 'f' suffix to 90+ float literals in array and divisions
- src/libWWW/Library/Implementation/HTInit.c: Added 'f' suffix to all float parameters
- src/libWWW/Library/Implementation/HTFile.c: Changed return 0.3 to 0.3f
- src/libWWW/Library/Implementation/HTFormat.c: Changed return -1e30 to -1e30f
- src/libIMG/bright.c: Added 'f' suffix to division operands
- src/libIMG/rle.c: Added explicit casts and 'f' suffix to float literals

Result: All 116 'double' to 'float' conversion warnings eliminated.

+++ src/libIMG/rle.c
@@ -166,23 +166,23 @@ Image* rleLoad(char* fullname, char* name, unsigned int verbose)
 
     debug("rleLoad: get_setup called ok\n");
     if (iflag == 1)              /* -i flag */
-        img_gam = 1.0 / img_gam; /* convert to display gamma */
+        img_gam = 1.0f / img_gam; /* convert to display gamma */
 
     /* If no image gamma on command line, check comments in file */
     if (!iflag) {
         char* v;
         if ((v = rle_getcom("image_gamma", &sv_globals)) != NULL) {
-            img_gam = atof(v);
+            img_gam = (float)atof(v);
             /* Protect against bogus information */
             if (img_gam == 0.0)
-                img_gam = 1.0;
+                img_gam = 1.0f;
             else
-                img_gam = 1.0 / img_gam; /* convert to display gamma */
+                img_gam = 1.0f / img_gam; /* convert to display gamma */
         } else if ((v = rle_getcom("display_gamma", &sv_globals)) != NULL) {
-            img_gam = atof(v);
+            img_gam = (float)atof(v);
             /* Protect */
             if (img_gam == 0.0)
-                img_gam = 1.0;
+                img_gam = 1.0f;
         }
     }
